+++ flink-doris-connector/src/main/java/org/apache/doris/flink/table/DorisRowDataJdbcLookupFunction.java
@@ -17,6 +17,7 @@
 
 package org.apache.doris.flink.table;
 
+import org.apache.doris.flink.lookup.LookupMetrics;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
